+++ src/Skeletest/Internal/Utils/Color.hs
@@ -2,8 +2,10 @@ module Skeletest.Internal.Utils.Color (
   green,
   red,
   yellow,
+  gray,
 ) where
 
+import Data.Colour.Names qualified as Color
 import Data.Text (Text)
 import Data.Text qualified as Text
 import System.Console.ANSI qualified as ANSI
@@ -19,3 +21,6 @@ red = withANSI [ANSI.SetColor ANSI.Foreground ANSI.Vivid ANSI.Red]
 
 yellow :: Text -> Text
 yellow = withANSI [ANSI.SetColor ANSI.Foreground ANSI.Vivid ANSI.Yellow]
+
+gray :: Text -> Text
+gray = withANSI [ANSI.SetRGBColor ANSI.Foreground Color.gray]